refactor(eth): move eth_log_from_event into a new events module
