+++ src/rpc/methods/eth/events/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2019-2026 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+//! The events of one executed tipset and their Ethereum form.
+
+mod logs;
+
+pub(super) use logs::eth_log_from_event;
